docs: changes to documentation
File Extraction

We removed the file because it was too large for uploading based on GitHub conditions. The new version of the file will be uploaded soon. 

The issue was detected from the following message:
                                       View Raw
(Sorry about that, but we can’t show files that are this big right now.)